backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
apscheduler==3.10.4
groq==0.9.0
python-multipart==0.0.9
email-validator==2.1.1
